+++ unikmer/cmd/inter.go
@@ -92,9 +92,6 @@ Tips:
 		var flag int
 
 		for i, file := range files {
-			if !firstFile && file == files[0] {
-				continue
-			}
 
 			if opt.Verbose {
 				log.Infof("processing file (%d/%d): %s", i+1, nfiles, file)