+++ unikmer/cmd/union.go
@@ -99,16 +99,11 @@ Tips:
 		var k int = -1
 		var canonical bool
 		var hasTaxid bool
-		var firstFile = true
 		var ok bool
 		var n int
 		var flag int
 		var nfiles = len(files)
 		for i, file := range files {
-			if !firstFile && file == files[0] {
-				continue
-			}
-
 			if opt.Verbose {
 				log.Infof("processing file (%d/%d): %s", i+1, nfiles, file)
 			}