+++ CHANGES.md
@@ -1,7 +1,8 @@
 - v0.9.0
     - `unikmer`: **new binary format supporting optional Taxids**.
     - deleted command: `unikmer subset`.
-    - new command: `unikmer head` for extracting the first N k-mers
+    - new command: `unikmer head` for extracting the first N k-mers.
+    - new command: `unikmer tsplit` for splitting k-mers according to taxid.
     - `unikmer grep`: support searching with taxids.
 - v0.8.0
     - `unikmer`: